Move oc commands to staging
