+++ pkg/oc/cli/cli.go
@@ -23,10 +23,14 @@ import (
 	"github.com/openshift/oc/pkg/cli/experimental/dockergc"
 	"github.com/openshift/oc/pkg/cli/expose"
 	"github.com/openshift/oc/pkg/cli/extract"
+	"github.com/openshift/oc/pkg/cli/kubectlwrappers"
+	"github.com/openshift/oc/pkg/cli/logout"
 	"github.com/openshift/oc/pkg/cli/logs"
 	"github.com/openshift/oc/pkg/cli/observe"
 	"github.com/openshift/oc/pkg/cli/options"
 	"github.com/openshift/oc/pkg/cli/recycle"
+	"github.com/openshift/oc/pkg/cli/rollback"
+	"github.com/openshift/oc/pkg/cli/rsh"
 	"github.com/openshift/oc/pkg/cli/secrets"
 	"github.com/openshift/oc/pkg/cli/serviceaccounts"
 	"github.com/openshift/oc/pkg/cli/whoami"
@@ -43,9 +47,7 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/idle"
 	"github.com/openshift/origin/pkg/oc/cli/image"
 	"github.com/openshift/origin/pkg/oc/cli/importimage"
-	"github.com/openshift/origin/pkg/oc/cli/kubectlwrappers"
 	"github.com/openshift/origin/pkg/oc/cli/login"
-	"github.com/openshift/origin/pkg/oc/cli/logout"
 	"github.com/openshift/origin/pkg/oc/cli/newapp"
 	"github.com/openshift/origin/pkg/oc/cli/newbuild"
 	"github.com/openshift/origin/pkg/oc/cli/policy"
@@ -54,9 +56,7 @@ import (
 	"github.com/openshift/origin/pkg/oc/cli/projects"
 	"github.com/openshift/origin/pkg/oc/cli/registry"
 	"github.com/openshift/origin/pkg/oc/cli/requestproject"
-	"github.com/openshift/origin/pkg/oc/cli/rollback"
 	"github.com/openshift/origin/pkg/oc/cli/rollout"
-	"github.com/openshift/origin/pkg/oc/cli/rsh"
 	"github.com/openshift/origin/pkg/oc/cli/rsync"
 	"github.com/openshift/origin/pkg/oc/cli/set"
 	"github.com/openshift/origin/pkg/oc/cli/startbuild"
