+++ staging/src/github.com/openshift/oc/pkg/cli/logout/logout.go
