+++ staging/src/github.com/openshift/oc/pkg/cli/rsh/rsh.go
